requirements.txt
assertpy==1.1
asyncpg==0.29.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytz==2024.1
randominfo==2.0.2
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3

selenium~=4.24.0
python-dotenv~=1.0.1